+++ scripts/validators/archive_node.sh
@@ -1,0 +1,1 @@
+./target/debug/ternoa --base-path ./tmp/archive --chain validator-net --name Archive_Node --port 30338 --ws-port 9959 --rpc-port 9978 --pruning=archive