+++ scripts/validators/validator_5.sh
@@ -1,0 +1,1 @@
+./target/debug/ternoa --base-path ./tmp/v5 --chain validator-net --validator --eve --name Eve_Node --port 30337 --ws-port 9948 --rpc-port 9937